remove getBlockDisplayInformation selector
