+++ docs/designers-developers/developers/data/data-core-blocks.md
@@ -6,10 +6,6 @@ Namespace: `core/blocks`.
 
 <!-- START TOKEN(Autogenerated selectors|../../../../packages/blocks/src/store/selectors.js) -->
 
-<a name="getBlockDisplayInformation" href="#getBlockDisplayInformation">#</a> **getBlockDisplayInformation**
-
-Undocumented declaration.
-
 <a name="getBlockStyles" href="#getBlockStyles">#</a> **getBlockStyles**
 
 Returns block styles by block name.
