+++ packages/block-editor/src/components/block-title/test/index.js
@@ -42,6 +42,7 @@ jest.mock( '@wordpress/blocks', () => {
 					return title;
 			}
 		},
+		getBlockVariations() {},
 	};
 } );
